Moduleify ServiceTypeDropdown.js
