+++ app/assets/javascripts/service_uploads/new_service_upload.jsx
@@ -1,8 +1,8 @@
 import Datepicker from '../student_profile/Datepicker.js';
+import ServiceTypeDropdown from './ServiceTypeDropdown.js';
 
 (function() {
   window.shared || (window.shared = {});
-  const ServiceTypeDropdown = window.shared.ServiceTypeDropdown;
 
   window.shared.NewServiceUpload = React.createClass({
